[cmyth] Handle MythTV priorities correctly
MythTV priorities can be significantly larger than the limits of a int8.
This change allows users with recordings that have priorities outside of the
int8 range to load recordings and use the MythTV PVR addon.

Without this, the addon will fail to populate the recordings list if there are
any recordings that had a priority outside the int8 range.

+++ lib/cmyth/include/cmyth/cmyth.h
@@ -1013,7 +1013,7 @@ extern uint32_t cmyth_proginfo_recordid(cmyth_proginfo_t prog);
  * \param prog proginfo handle
  * \return int32_t
  */
-extern int8_t cmyth_proginfo_priority(cmyth_proginfo_t prog);
+extern int32_t cmyth_proginfo_priority(cmyth_proginfo_t prog);
 
 /**
  * Retrieve the critics rating (number of stars) of a program.
